api/pyproject.toml
[tool.poetry]
name = "bracket-app"
version = "0.1.0"
description = ""
authors = ["Grzegorz Dlugokecki <grzesiek.dlugokecki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
sqlmodel = "^0.0.31"
pydantic-settings = "^2.12.0"
alembic = "^1.17.2"
asyncpg = "^0.31.0"
pwdlib = {extras = ["argon2"], version = "^0.3.0"}
pyjwt = "^2.10.1"
pydantic = {extras = ["email"], version = "^2.12.5"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 120
